partilce-wave duality

+++ shp/axis.cpp
@@ -22,11 +22,11 @@
 
 namespace shp {
 
-Axis::Axis() : name(""), scaling(0.0) {
+Axis::Axis() : name(""), scaling(0.0f) {
 
 }
 
-Axis::Axis(std::string name, double scaling)
+Axis::Axis(std::string name, float scaling)
         : name(name), scaling(scaling) {
 
 }
